use variables for kvd home and kvcli home, check for errors

+++ rest_test/run_all_tests_from_make.sh
@@ -11,13 +11,22 @@ faucet="chief access utility giant burger winner jar false naive mobile often pe
 # address: kava1ls82zzghsx0exkpr52m8vht5jqs3un0ceysshz
 # address: kavavaloper1ls82zzghsx0exkpr52m8vht5jqs3un0c5j2c04
 
+# variables for home directories for kvd and kvcli
+kvdHome=/tmp/kvdHome
+kvcliHome=/tmp/kvcliHome
+
 # Remove any existing data directory
-rm -rf ~/kavatmp
+rm -rf $kvdHome
+rm -rf $kvcliHome
+
+# make the directories
+mkdir /tmp/kvdHome
+mkdir /tmp/kvcliHome
 
 # create validator key
-printf "$password\n$validatorMnemonic\n" | kvcli keys add vlad --recover --home ~/kavatmp
+printf "$password\n$validatorMnemonic\n" | kvcli keys add vlad --recover --home $kvcliHome
 # create faucet key
-printf "$password\n$faucet\n" | kvcli --home ~/kavatmp keys add faucet --recover --home ~/kavatmp
+printf "$password\n$faucet\n" | kvcli --home $kvcliHome keys add faucet --recover --home $kvcliHome
 
 # function used to show that it is still loading
 showLoading() {
@@ -41,41 +50,42 @@ showLoading() {
   echo "$loadingText...finished"
 }
 
-# build the go setup test file
-rm -f rest_test/setuptest
-go build rest_test/setup/setuptest.go & showLoading "Building go test file, please wait"
-
 # Create new data directory
 {
-kvd --home ~/kavatmp init --chain-id=testing vlad # doesn't need to be the same as the validator
+kvd --home $kvdHome init --chain-id=testing vlad # doesn't need to be the same as the validator
 } > /dev/null 2>&1
-kvcli --home ~/kavatmp config chain-id testing # or set trust-node true
+kvcli --home $kvcliHome config chain-id testing # or set trust-node true
 
 # add validator account to genesis
-kvd --home ~/kavatmp add-genesis-account $(kvcli --home ~/kavatmp keys show vlad -a) 10000000000000stake
+kvd --home $kvdHome add-genesis-account $(kvcli --home $kvcliHome keys show vlad -a) 10000000000000stake
 # add faucet account to genesis
-kvd --home ~/kavatmp add-genesis-account $(kvcli --home ~/kavatmp keys show faucet -a) 10000000000000stake,1000000000000xrp,100000000000btc
+kvd --home $kvdHome add-genesis-account $(kvcli --home $kvcliHome keys show faucet -a) 10000000000000stake,1000000000000xrp,100000000000btc
 
 # Create a delegation tx for the validator and add to genesis
-printf "$password\n" | kvd --home ~/kavatmp gentx --name vlad --home-client ~/kavatmp
+printf "$password\n" | kvd --home $kvdHome gentx --name vlad --home-client $kvcliHome
 {
-kvd --home ~/kavatmp collect-gentxs
+kvd --home $kvdHome collect-gentxs
 } > /dev/null 2>&1
 
 # start the blockchain in the background, wait until it starts making blocks
 {
-kvd start --home ~/kavatmp & kvdPid="$!"
+kvd start --home $kvdHome & kvdPid="$!"
 } > /dev/null 2>&1
 
 printf "\n"
 sleep 10 & showLoading "Starting rest server, please wait"
 # start the rest server. Use ./stopchain.sh to stop both rest server and the blockchain
 {
-kvcli rest-server --laddr tcp://127.0.0.1:1317 --chain-id=testing --home ~/kavatmp & kvcliPid="$!"
+kvcli rest-server --laddr tcp://127.0.0.1:1317 --chain-id=testing --home $kvcliHome & kvcliPid="$!"
 } > /dev/null 2>&1
 printf "\n"
 sleep 10 & showLoading "Preparing blockchain setup transactions, please wait"
 printf "\n"
+
+# build the go setup test file
+rm -f rest_test/setuptest
+go build rest_test/setup/setuptest.go & showLoading "Building go test file, please wait"
+
 # run the go code to send transactions to the chain and set it up correctly
 ./setuptest & showLoading "Sending messages to blockchain"
 printf "\n"
@@ -99,13 +109,17 @@ then
   # check that all the tests passed (ie zero failing)
   if [[ $(cat output | grep "0 failing") ]]
   then
-    echo "Success"
-    rm setuptest & showLoading "Cleaning up go binary"
-    # kill the kvd and kvcli processes (blockchain and rest api)
-    pgrep kvd | xargs kill
-    pgrep kvcli | xargs kill & showLoading "Stopping blockchain"
-    rm -f output
-    exit 0
+    # check for no errors
+    if [[ $(cat output | grep "0 errors") ]]
+    then
+      echo "Success"
+      rm setuptest & showLoading "Cleaning up go binary"
+      # kill the kvd and kvcli processes (blockchain and rest api)
+      pgrep kvd | xargs kill
+      pgrep kvcli | xargs kill & showLoading "Stopping blockchain"
+      rm -f output
+      exit 0
+    fi
   fi
 fi
